# server/swagger.yaml
openapi: 3.0.0
info:
  title: Agentic Schedule API
  version: 1.0.0
  description: API documentation for Contracts, Staff, Skills, DayOffRequests, and ShiftOffRequests
paths:
  /contracts:
    post:
      summary: Create one or multiple contracts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Contract'
                - type: array
                  items:
                    $ref: '#/components/schemas/Contract'
            examples:
              SingleContract:
                summary: Single contract object
                value:
                  ContractID: 1
                  Description: "Sample contract"
                  SingleAssignmentPerDay: true
                  MaxNumAssignments: 5
                  MinNumAssignments: 1
                  MaxConsecutiveWorkingDays: 6
                  MinConsecutiveWorkingDays: 2
                  MaxConsecutiveFreeDays: 3
                  MinConsecutiveFreeDays: 1
                  MaxConsecutiveWorkingWeekends: 2
                  MinConsecutiveWorkingWeekends: 1
                  MaxWorkingWeekendsInFourWeeks: 3
                  WeekendDefinition: "Sat-Sun"
                  CompleteWeekends: true
                  IdenticalShiftTypesDuringWeekend: false
                  NoNightShiftBeforeFreeWeekend: false
                  AlternativeSkillCategory: false
                  UnwantedPatterns: ""
              MultipleContracts:
                summary: List of contract objects
                value:
                  - ContractID: 2
                    Description: "Contract 2"
                    SingleAssignmentPerDay: false
                    MaxNumAssignments: 4
                    MinNumAssignments: 2
                    MaxConsecutiveWorkingDays: 5
                    MinConsecutiveWorkingDays: 2
                    MaxConsecutiveFreeDays: 2
                    MinConsecutiveFreeDays: 1
                    MaxConsecutiveWorkingWeekends: 1
                    MinConsecutiveWorkingWeekends: 1
                    MaxWorkingWeekendsInFourWeeks: 2
                    WeekendDefinition: "Sat-Sun"
                    CompleteWeekends: false
                    IdenticalShiftTypesDuringWeekend: false
                    NoNightShiftBeforeFreeWeekend: false
                    AlternativeSkillCategory: false
                    UnwantedPatterns: ""
                  - ContractID: 3
                    Description: "Contract 3"
                    SingleAssignmentPerDay: true
                    MaxNumAssignments: 6
                    MinNumAssignments: 3
                    MaxConsecutiveWorkingDays: 7
                    MinConsecutiveWorkingDays: 2
                    MaxConsecutiveFreeDays: 3
                    MinConsecutiveFreeDays: 1
                    MaxConsecutiveWorkingWeekends: 2
                    MinConsecutiveWorkingWeekends: 1
                    MaxWorkingWeekendsInFourWeeks: 4
                    WeekendDefinition: "Fri-Sat"
                    CompleteWeekends: true
                    IdenticalShiftTypesDuringWeekend: true
                    NoNightShiftBeforeFreeWeekend: true
                    AlternativeSkillCategory: true
                    UnwantedPatterns: ""
      responses:
        '201':
          description: Contracts created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 201
                  success_message:
                    type: string
                    example: Contracts created successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      contractIds:
                        type: array
                        items:
                          type: integer
                        example: [1, 2, 3]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    get:
      summary: Get all contracts
      responses:
        '200':
          description: List of contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardContractsListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    delete:
      summary: Delete contracts by list of ContractID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contractIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
      responses:
        '200':
          description: Contracts deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Contracts deleted successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                        example: 3
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
...
  /dayoffrequests:
    post:
      summary: Create one or multiple day off requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/DayOffRequest'
                - type: array
                  items:
                    $ref: '#/components/schemas/DayOffRequest'
      responses:
        '201':
          description: Day off requests created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 201
                  success_message:
                    type: string
                    example: Day off requests created successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      requestIds:
                        type: array
                        items:
                          type: integer
                        example: [1, 2, 3]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    get:
      summary: Get all day off requests
      responses:
        '200':
          description: List of day off requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Day off requests fetched successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      dayOffRequests:
                        type: array
                        items:
                          $ref: '#/components/schemas/DayOffRequest'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    delete:
      summary: Delete day off requests by list of RequestID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requestIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
      responses:
        '200':
          description: Day off requests deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Day off requests deleted successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                        example: 2
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
...
components:
  schemas:
    Contract:
      type: object
      properties:
        ContractID:
          type: integer
        Description:
          type: string
        SingleAssignmentPerDay:
          type: boolean
        MaxNumAssignments:
          type: integer
        MinNumAssignments:
          type: integer
        MaxConsecutiveWorkingDays:
          type: integer
        MinConsecutiveWorkingDays:
          type: integer
        MaxConsecutiveFreeDays:
          type: integer
        MinConsecutiveFreeDays:
          type: integer
        MaxConsecutiveWorkingWeekends:
          type: integer
        MinConsecutiveWorkingWeekends:
          type: integer
        MaxWorkingWeekendsInFourWeeks:
          type: integer
        WeekendDefinition:
          type: string
        CompleteWeekends:
          type: boolean
        IdenticalShiftTypesDuringWeekend:
          type: boolean
        NoNightShiftBeforeFreeWeekend:
          type: boolean
        AlternativeSkillCategory:
          type: boolean
        UnwantedPatterns:
          type: string
...
    DayOffRequest:
      type: object
      properties:
        employeeId:
          type: integer
        requestDate:
          type: string
          format: date
        requestTypeId:
          type: integer
...
    StandardErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        statusCode:
          type: integer
          example: 400
        success_message:
          type: string
          example: ""
        error_message:
          type: string
          example: "Invalid data"
        payload:
          type: object
          example: {}
    StandardContractsListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        statusCode:
          type: integer
          example: 200
        success_message:
          type: string
          example: Contracts fetched successfully
        error_message:
          type: string
          example: ""
        payload:
          type: object
          properties:
            contracts:
              type: array
              items:
                $ref: '#/components/schemas/Contract'
...